# KIALI-1081 Add concurrence for each virtual service validation