--- a/services/business/checkers/virtual_service_checker.go
+++ b/services/business/checkers/virtual_service_checker.go
@@ -1,6 +1,8 @@
 package checkers
 
 import (
+	"sync"
+
 	"github.com/kiali/kiali/kubernetes"
 	"github.com/kiali/kiali/services/business/checkers/virtual_services"
 	"github.com/kiali/kiali/services/models"
@@ -30,11 +32,24 @@ func (in VirtualServiceChecker) Check() models.IstioValidations {
 // Runs individual checks for each virtual service
 func (in VirtualServiceChecker) runIndividualChecks() models.IstioValidations {
 	validations := models.IstioValidations{}
+	validationsChan := make(chan models.IstioValidations)
+
+	var wg sync.WaitGroup
+	wg.Add(len(in.VirtualService))
 
 	for _, virtualService := range in.VirtualService {
-		in.runChecks(virtualService, validations)
+		go in.runChecks(virtualService, validationsChan, &wg)
 	}
 
+	go func() {
+		wg.Wait()
+		// Closing the channel stop the range loop below
+		close(validationsChan)
+	}()
+
+	for v := range validationsChan {
+		validations.MergeValidations(v)
+	}
 	return validations
 }
 
@@ -44,7 +59,11 @@ func (in VirtualServiceChecker) runGroupChecks() models.IstioValidations {
 }
 
 // runChecks runs all the individual checks for a single virtual service and appends the result into validations.
-func (in VirtualServiceChecker) runChecks(virtualService kubernetes.IstioObject, validations models.IstioValidations) {
+func (in VirtualServiceChecker) runChecks(virtualService kubernetes.IstioObject,
+	validationChan chan models.IstioValidations, wg *sync.WaitGroup) {
+
+	defer (*wg).Done()
+
 	virtualServiceName := virtualService.GetObjectMeta().Name
 	key := models.IstioValidationKey{Name: virtualServiceName, ObjectType: virtualCheckerType}
 	rrValidation := &models.IstioValidation{
@@ -64,5 +83,5 @@ func (in VirtualServiceChecker) runChecks(virtualService kubernetes.IstioObject,
 		rrValidation.Valid = rrValidation.Valid && validChecker
 	}
 
-	validations.MergeValidations(models.IstioValidations{key: rrValidation})
+	validationChan <- models.IstioValidations{key: rrValidation}
 }