JIRA: [KIALI-1081](https://issues.jboss.org/browse/KIALI-1081)

I have done a little benchmark and the concurrent version had better performance over the sequential one. Benchmark aimed to discard over-concurrency.