@jotak 2018-07-06
@xeviknal , I've run some benchmarks and find the opposite of yours: synchronous code is almost twice faster than async. So I don't know where it went wrong. But I think the comment I had here [1] and the previous work done by Simon were right.

If nothing is wrong in my benchmark, still, it's not a big deal. But if we can simplify the code **and** have it run faster, why wouldn't do it?

Note, about go benchmarking: https://dave.cheney.net/2013/06/30/how-to-write-benchmarks-in-go
My benchmark is:

```go
var gValidations models.IstioValidations

func BenchmarkVirtualServiceMultipleIstioObjects(b *testing.B) {
	var v models.IstioValidations
	for n := 0; n < b.N; n++ {
		// Setup mocks
		destinationList := []kubernetes.IstioObject{
			fakeDestinationRule("reviews"),
		}

		virtualServiceChecker := VirtualServiceChecker{"bookinfo",
			destinationList, fakeVirtualServiceMultipleIstioObjects()}

		v = virtualServiceChecker.Check()
	}
	gValidations = v
}
```

Result before this PR:
```
First run: BenchmarkVirtualServiceMultipleIstioObjects-8   	  100000	     12210 ns/op	    9567 B/op	      95 allocs/op
Second run: BenchmarkVirtualServiceMultipleIstioObjects-8   	  100000	     12287 ns/op	    9567 B/op	      95 allocs/op
Third run: BenchmarkVirtualServiceMultipleIstioObjects-8   	  100000	     12272 ns/op	    9567 B/op	      95 allocs/op
```

Result after this PR:
```
First run: BenchmarkVirtualServiceMultipleIstioObjects-8   	  100000	     22009 ns/op	    9687 B/op	      97 allocs/op
Second run: BenchmarkVirtualServiceMultipleIstioObjects-8   	  100000	     20324 ns/op	    9687 B/op	      97 allocs/op
Third run: BenchmarkVirtualServiceMultipleIstioObjects-8   	  100000	     20378 ns/op	    9686 B/op	      97 allocs/op
```

So IMO it could be reverted. Or at least, if you don't want to, I think from now on we should be a bit more careful while adding goroutines.

[1] https://issues.jboss.org/browse/KIALI-1081?focusedCommentId=13599938&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-13599938