@lucasponce 2018-07-06
It is good to know, and nothing wrong to add a JIRA to refine this, but at low priority I guess.
In terms of overall behaviour from the app, I guess there are other areas open with more priority (in terms of graph, pagination or health calculation).