@jotak 2018-07-06
@lucasponce you're right that there's tasks with higher priority. As well as there were tasks with higher priority than this one (KIALI-1081) which was not important **and** wrong at the same time :-(

PS: a `git revert` is a 1-minute task or so.