@lucasponce 2018-07-06
@jotak KIALI-1081 was to unify both designs in terms of code consistenty.
So it was not wrong.
There was a long discussion on https://github.com/kiali/kiali/pull/232 to adopt a model for the checkers.

So, unify the code to follow a single design was the goal.
If there is something broken, good to change, if it is an improvement, it is good as well.
But, we should think in global.

A JIRA to change the design in all the code is not bad, but I don't want to to a git revert of a single area to maintain two styles, and then in one week we will question these things.

If this is a minor change, a JIRA to change the concurrency in all checkers is good, but again, let's think in perspective.