@jotak 2018-07-06
Well, I could agree if it was really a matter of style, but adding threads is not just a matter of style, it's really more than that. I don't subscribe to this "all-or-nothing" view in such case, that's all I can say.