@lucasponce 2018-07-06
The point is that:
- We had a discussion in #232 for the same topic.
- New additions should follow same design.
- We have detected design can be improved (that's good).
- Then we change the design in all checkers to unify things.

But I am -1 to revert partially changes.

As I commented a JIRA to do the unification is fine with me, but I wouldn't revert this and maintain both designs, I prefer a global change in all the checkers.

I am not arguing that we should maintain the concurrency. I am trying to maintain some consistency and propagate changes accordingly.

Please, feel free to create the JIRA.